hex-literal v0.3
