+++ hex-literal/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hex-literal"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["RustCrypto Developers"]
 license = "MIT OR Apache-2.0"
 description = "Procedural macro for converting hexadecimal string to byte array at compile time."
@@ -9,6 +9,5 @@ repository = "https://github.com/RustCrypto/utils"
 keywords = ["hex", "proc-macro", "literals"]
 edition = "2018"
 
-[dependencies]
-proc-macro-hack = "0.5"
-hex-literal-impl = { version = "0.2", path = "./hex-literal-impl" }
+[lib]
+proc-macro = true
