--- hex-literal/hex-literal-impl/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "hex-literal-impl"
-version = "0.2.2"
-authors = ["RustCrypto Developers"]
-license = "MIT OR Apache-2.0"
-description = "Internal implementation of the hex-literal crate"
-repository = "https://github.com/RustCrypto/utils"
-edition = "2018"
-
-[dependencies]
-proc-macro-hack = "0.5"
-
-[lib]
-proc-macro = true
-
